requirements.txt
flask>=3.0
python-docx>=1.1
openpyxl>=3.1
pypdf>=4.0
